Update 2_categorize_app_basel.py
Renamed playsound() to play_sound() to avoid name collision with playsound package. Modified this function to try pydub first and use playsound as a backup, then remember which one to keep using.

Changed dropdown menu options. They are no longer hardcoded in, but now read from a textfile. The script looks for text files with certain names to import the dropdown options